+++ htmlresources/lib/azgridview/tablefieldsgridview.js
@@ -1,8 +1,8 @@
 class TableFieldsGridView extends AZGridView {
 
-    constructor() {
-        super('fieldsgrid', 
-            [{name: 'id', caption: 'Id', style: 'width: 80px;'},
+    constructor(showPK) {
+        let fields = [
+            {name: 'id', caption: 'Id', style: 'width: 80px;'},
             {name: 'name', caption: 'Name', style: 'width: 50%;'},
             //{name: 'caption', caption: 'Caption', style: 'width: 30%'},
             {name: 'dataType', caption: 'Data Type', style: 'width: 20%', autocomplete: [                
@@ -12,7 +12,14 @@ class TableFieldsGridView extends AZGridView {
             {name: 'length', caption: 'Length', style: 'width:100px'},
             {name: 'dataClassification', caption: 'Data Classification', style: 'width: 20%', autocomplete: [
                 'AccountData', 'CustomerContent', 'EndUserIdentifiableInformation', 'EndUserPseudonymousIdentifiers',
-                'OrganizationIdentifiableInformation', 'SystemMetadata', 'ToBeClassified']}],
+                'OrganizationIdentifiableInformation', 'SystemMetadata', 'ToBeClassified']}];
+        
+        if (showPK)
+            fields.unshift(
+                {name: 'pk', caption: 'PK', style: 'width: 32px', type: 'boolean'}
+            );
+
+        super('fieldsgrid', fields,
             undefined, undefined, 'Loading data...', true);
     }
 }