+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 2.0.6
  - Issue #153 - Hide content of custom editor and show warning if document cannot be parsed
-
+ - Issue #162 - Table Wizard: Mark PK fields
+ - Boolean column type support added to the data grid
+ 
 ## 2.0.5
 
  - Issue #131 - "AL Outline" panel is blank - legacy Nav2018 version of the language server added